+++ crates/nvim/src/nvim/drawline/columns.rs
@@ -64,7 +64,7 @@ impl WinLineVars {
         // SAFETY: the caller's buffer, and `off` is kept under the view width
         // by the loop condition.
         unsafe {
-            let end = text.add(len as usize);
+            let end = text.add(len);
             let mut ptr = text;
             while ptr < end && self.off < (*wp).w_view_width {
                 let cells = line_putchar(
